Configure webpack-dev-server to act as a full-fledged dev server
webpack-dev-server can do request proxying and exposes the underlying
Express app for configuration.
As a result it can do the same as server.js

It is now a matter of taste if one wants to configure Express through WDS
or the other way around.

`npm run dev-server` starts server.js as until now.
`npm run dev-start` starts WDS which is now fully usable.